% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donutsk-package.R
\docType{package}
\name{donutsk-package}
\alias{donutsk}
\alias{donutsk-package}
\title{donutsk: Construct advanced donut charts}
\description{
The Donutsk package allows you to build donut/pie charts with ggplot2 layer by layer, exploiting the advantages of polar symmetry.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dkibalnikov/donutsk}
  \item Report bugs at \url{https://github.com/dkibalnikov/donutsk/issues}
}

}
\author{
\strong{Maintainer}: Dmitry Kibalnikov \email{d.kibalnikov@gmail.com}

}
\keyword{internal}
