% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDP}
\alias{GDP}
\title{The World Bank GDP, PPP (current international $)}
\format{
\subsection{\code{GDP}}{

A data frame with 6,004 rows and 5 columns:
\describe{
\item{date}{Year}
\item{country}{Country name}
\item{region}{Region hierarchy}
\item{region_ISO}{3 letter ISO region codes}
\item{GDP}{GDP, PPP (current international $)}
...
}
}
}
\source{
\url{https://data.worldbank.org/indicator/NY.GDP.MKTP.PP.CD}
}
\usage{
GDP
}
\description{
\itemize{
\item A pre-processed subset of GDP data from the World Bank
\item GDP, PPP means gross domestic product based on purchasing power parity
\item \emph{current international $} means actual (not adjusted to inflation) US dollars
}
}
\keyword{datasets}
