% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donutsk-package.R
\docType{package}
\name{donutsk-package}
\alias{donutsk}
\alias{donutsk-package}
\title{donutsk: Construct Advanced Donut Charts}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Build donut/pie charts with 'ggplot2' layer by layer, exploiting the advantages of polar symmetry. Leverage layouts to distribute labels effectively. Connect labels to donut segments using pins. Streamline annotation and highlighting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dkibalnikov/donutsk}
  \item \url{https://dkibalnikov.github.io/donutsk/}
  \item Report bugs at \url{https://github.com/dkibalnikov/donutsk/issues}
}

}
\author{
\strong{Maintainer}: Dmitry Kibalnikov \email{d.kibalnikov@gmail.com} [copyright holder]

}
\keyword{internal}
